functions/azure-functions-todo-notifications/requirements.txt
azure-functions
azure-cosmos
sendgrid
python-dateutil
requests